+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [UNRELEASED] 
+### Added 
+* magnet command to generate magnet URI's from a bittorrent metafile.
+
 ## [v0.2.0] - 2021-01-25
 ### Added
 * Add experimental windows support and installer.